requirements.txt
python-telegram-bot[job-queue]==20.7
requests==2.31.0
jdatetime==4.1.1
python-dotenv==1.0.0
aiosqlite==0.19.0
asyncio==3.4.3
APScheduler==3.10.4 